requirements.txt
# Core dependencies
openai==1.30.0
python-telegram-bot==20.7
python-dotenv==1.0.0

# Audio processing dependencies
SpeechRecognition==3.10.0
gTTS==2.3.2
pydub==0.25.1

# Optional audio input dependency (for microphone support)
# PyAudio==0.2.14

# Additional system dependencies (install separately):
# - ffmpeg (required for pydub audio conversion)
# - portaudio19-dev (Linux) / portaudio (macOS) - required for PyAudio
# - python3-dev (Linux) - required for PyAudio compilation
